refactor: split java adapter modules
